requirements.txt
Django==1.10
django-braces==1.13.0
django-ckeditor==5.2.1
django-cleanup==4.0.0
django-crispy-forms==1.6.1
django-model-utils==3.2.0
django-recaptcha==1.2.1
django-star-ratings==0.8.0
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
pytz==2019.3
sqlparse==0.3.0
swapper==1.1.1
